ci: ensure that the builder container is healthy
GitHub Actions sometimes fails due to unclear reasons with the following
error:

```
$ ./build-exec bundle config set --local without check
There was an error while trying to write to `/home/me/build/.bundle/config`. It
is likely that you need to grant write permissions for that path.
```

This commit tries to fix this by waiting for the builder container