build: Remove `--gemfile` option from `rbwasm` command line
This option was used to specify the path to the `Gemfile` by overriding
gemfile path after launching rbwasm command through Bundler. But it does
not work because Bundler caches a lot of information and cannot
invalidate all of them in-process. So we should cd to the directory
containing the Gemfile and run the command there to avoid the need to
switch the Gemfile path.

This issue was revealed when trying to install remote gems, which was
reuired to work around installation issue around bundled power_assert
gem in the Ruby build system.