Catch exceptions thrown during `rbVmBugreport`
The `rbVmBugreport` operation itself can throw exceptions because the VM
state might be in a bad shape at this point due to unexpected
non-RbError exceptions. And if it does, it forgets to report the
original exception that actually caused the bad state.
This change catches exceptions thrown during `rbVmBugreport` and just
logs them instead of rethrowing them to ensure that the original
exception is not lost.