Fix ruby cache key calculation used for npm packaging
The `RUBY_WASM_ROOT` environment variable is not set when calculating
ruby cache keys for npm packages and rbwasm assumes that the
build_manifest.json file is placed in the cwd (npm package directory).
But the build_manifest.json file is placed in the repository root
so the cache key calculation didn't respect the pinned ruby revision.