Fix missing Asyncify-loop on exposed entry points
Some of operations that *may* trigger Asyncify-unwind were not
surrounded by the Asyncify-loop. Without the loop, the unwinding
would not be stopped at the boundary of the guest function, and the
continuation would be lost. Typically, unwinding for scanning GC root
from Wasm locals were not rewound properly, then the GC would miss
some objects and cause a crash.